requirements.txt
streamlit
langchain
python-dotenv
openai
langchain_community
langchain-openai
langchain-google-genai
reportlab
jinja2
pdfkit
fpdf
fastapi
uvicorn
email-validator
reportlab